+++ release_notes.md
@@ -2,7 +2,8 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- Removing host level retry specification (#7870)
 
 **Release sprint:** Sprint 113
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+113%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+113%22+label%3Afeature+is%3Aclosed) ]
+- Update PowerShell Worker 7.0 to 4.0.1273 [Release Note](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.1273)
+- Update PowerShell Worker 7.2 to 4.0.1274 [Release Note](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.1274)